Fix SUPER BUG in quadrotor position controller
Anti-climax, this commit is NOT REALLY A BUG in the position controller
per se, but rather a problem with regards to the rates of which the
position controller is operating. The pose callback from the prototype
quadrotor simulation is at 1000Hz, when we feed the pose data into the
position controller, the dominant error comes from the integral term
which is essentiall FLOAT_MAX, making it difficult to have small error
signals. To fix this problem, I simply set the integral term to zero,
however this is not an ideal solution, either:

- Make position controller run at a much slower rate than 1000hz forced
  by the pose callback (force a ros::Rate::sleep() inside the pose
  callback?)
- Add integral term safe-guards

The first option is preferrable.