Add ability to recurse down config key
Previously ConfigParser assumes the yaml config has only 1 level, e.g.

    variable:
        x: 1
        y: 2
        z: 3

However, obviously there are cases where we would like more than 1 level
of nested yaml config. This commit allows the user to specify the nested
key as like so:

    level1:
        level2:
            x: 1

    ConfigParser.addParam("level1.leve2.x");

As seen above we use the dot-notation to denote the nested level in the
yaml config.